backend/requirements.txt
fastapi==0.104.1
uvicorn==0.38.0
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
mongoengine==0.29.1
pymongo==4.15.4
python-multipart==0.0.20
minio==7.2.20
dnspython==2.8.0
slowapi==0.1.9
